chara_fur: makeFurTex FurRandScale signed int->float conversion (xoris, census-verified; unblocked by named temps) (78.60 -> 79.53, unit +0.22)

+++ src/chara_fur.cpp
@@ -704,7 +704,7 @@ static inline unsigned int FurRandNext()
 
 static inline float FurRandScale()
 {
-	return kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(FurRandNext()) + kCharaFurViewDepthThreshold);
+	return kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(static_cast<int>(FurRandNext())) + kCharaFurViewDepthThreshold);
 }
 
 static inline CColor FurNoiseColor(const CColor& base, const CColor& noise, float scale)
